finished the project

+++ src/main/kotlin/com/api/kotlin/repository/EmployeeRepository.kt
@@ -3,5 +3,5 @@ package com.api.kotlin.repository
 import com.api.kotlin.model.Employee
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface EmployeeRepository : JpaRepository<Employee, Int> {
+interface EmployeeRepository : JpaRepository<Employee, Long> {
 }
